+++ src/jobs/TriggerIntegration.php
@@ -7,6 +7,7 @@ use verbb\formie\elements\Submission;
 use verbb\formie\models\IntegrationResponse;
 
 use Craft;
+use craft\db\Query;
 use craft\helpers\Json;
 
 use Exception;
@@ -41,6 +42,13 @@ class TriggerIntegration extends BaseJob
             // Pass a reference of this class to the integration, to assist with debugging.
             // Set with a private variable, so it doesn't appear in the queue job data which would be mayhem.
             $this->integration->setQueueJob($this);
+
+            // Re-fetch the token in case it's been changed (as we store a reference to the entire integration in the job)
+            $this->integration->tokenId = (new Query())
+                ->select('tokenId')
+                ->from(['{{%formie_integrations}}'])
+                ->where(['uid' => $this->integration->uid])
+                ->scalar();
 
             // Ensure we set the correct language for a potential CLI request
             Craft::$app->language = $submission->getSite()->language;