+++ CHANGELOG.md
@@ -452,6 +452,20 @@
 - `Field::name` attribute has been deprecated. Use `Field::label` instead.
 - `Field::inputHtml()` method has been deprecated. Use `Field::cpInputHtml()` instead.
 
+## 2.1.16 - 2024-05-27
+
+### Added
+- Phone fields now include `countryCode` and `countryName` in their value when the value is JSON.
+
+### Changed
+- Removed unused `e.target` from Repeater `addRow()` JS function.
+- Updated English translations.
+- Updated reCAPTCHA Enterprise’s Secret API Key plugin setting.
+
+### Fixed
+- Fixed Entry element integrations not using their section’s default entry status when `enabled` wasn’t mapped.
+- Fixed country-enabled Phone fields not having their generated value set correctly.
+
 ## 2.1.15 - 2024-05-20
 
 ### Fixed
@@ -1891,6 +1905,11 @@
 - Removed `enableGatsbyCompatibility` plugin setting, as it's no longer needed.
 - Removed `forms`, `form` and `formCount` from GraphQL queries. Please use `formieForms`, `formieForm` and `formieFormCount`.
 - Removed `submissions`, `submission` and `submissionCount` from GraphQL queries. Please use `formieSubmissions`, `formieSubmission` and `formieSubmissionCount`.
+
+## 1.6.42 - 2024-05-20
+
+### Fixed
+- Fix an error when triggering an integration where the `tokenId` has become invalid.
 
 ## 1.6.41 - 2024-04-18